client/requirements.txt
grpcio
grpcio-tools
#tensorflow
#keras
pandas
